added model_columns unit tests
Moved parent_child unit tests to folder.
Added unit tests for child columns.
Removed split_annotation_fields
