+++ src/pandas_to_pydantic/__init__.py
@@ -1,6 +1,4 @@
 from pandas_to_pydantic.annotation_utils import (  # noqa: F401
-    expand_annotation,
     get_model_columns,
-    split_annotation_fields,
 )
 from pandas_to_pydantic.to_pydantic import dataframe_to_pydantic, get_root_list, serialize_dataframe  # noqa: F401
