+++ src/pandas_to_pydantic/annotation_utils.py
@@ -4,7 +4,6 @@ from pydantic import BaseModel
 from pydantic._internal._model_construction import ModelMetaclass
 
 
-# TODO consider renaming fields, both are child models
 class ModelColumns(BaseModel):
     name: str
     id_column: str | None
@@ -16,6 +15,7 @@ class ModelColumns(BaseModel):
 def get_model_columns(
     model: ModelMetaclass, id_column_map: dict[str, str] | None = None, name: str | None = None
 ) -> ModelColumns:
+    # TODO consider returning field name
     if not model.__base__ == BaseModel:
         error_message = f"{model} is not a BaseModel"
         raise TypeError(error_message)
@@ -25,7 +25,6 @@ def get_model_columns(
     if name is None:
         name = model.__name__
 
-    # TODO name being different than class name is a bit unintuitive
     id_column = id_column_map.get(name)
     annotations = model.__annotations__
 
@@ -40,7 +39,7 @@ def get_model_columns(
                 list_columns.append(get_model_columns(field_type.__args__[0], id_column_map, field_name))
         elif isinstance(field_type, ModelMetaclass):
             if field_type.__base__ == BaseModel:
-                child_columns.append(get_model_columns(field_type, id_column_map))
+                child_columns.append(get_model_columns(field_type, id_column_map, field_name))
         else:
             base_columns.append(field_name)
 
@@ -85,16 +84,3 @@ def expand_annotation(model: ModelMetaclass) -> dict:
                 annotations[field_name] = expand_annotation(field_type)
 
     return annotations
-
-
-def split_annotation_fields(annotation: dict) -> dict[str, list[str]]:
-    base_fields = []
-    list_fields = []
-
-    for field_name, field_type in annotation.items():
-        if isinstance(field_type, list):
-            list_fields.append(field_name)
-        else:
-            base_fields.append(field_name)
-
-    return {"base": base_fields, "list": list_fields}
